chore(module-resolution) Refactored to make the logic more closely follow the pseudocode
